V 2.1

+++ IntVector.h
@@ -6,8 +6,8 @@
 typedef struct 
 {
 	int *data;
-	int size;
-	int capacity;
+	size_t size;
+	size_t capacity;
 } IntVector;
 
 IntVector *int_vector_new(size_t initial_capacity);
